+++ R/latest.R
@@ -10,6 +10,9 @@
 #' @param parameter  Limit to only a certain parameter (valid values are 'pm25', 'pm10', 'so2', 'no2', 'o3', 'co' and 'bc').
 #' If no parameter is given, all parameters are retrieved.
 #' @param has_geo has_geo Filter out items that have or do not have geographic information.
+#' @param latitude Latitude of the center point (lat, lon) used to get measurements within a certain area.
+#' @param longitude Longitude of the center point (lat, lon) used to get measurements within a certain area.
+#' @param radius Radius (in meters) used to get measurements within a certain area, must be used with latitude and longitude
 #' @param limit Change the number of results returned, max is 1000.
 #' @param page The page of the results to query. This can be useful if e.g. there are 2000 measurements, then first use page=1 and page=2 with limit=100 to get all measurements for your query.
 
@@ -23,10 +26,13 @@
 #'   one day there is a city with the same name in another country.
 #' @examples
 #' \dontrun{
-#' aq_latest(country='IN', city='Chennai')
+#' latest_chennai <- aq_latest(country='IN', city='Chennai')
+#' latest_chennai
+#' attr(latest_chennai, "meta")
+#' attr(latest_chennai, "timestamp")
 #' aq_latest(parameter='co')
 #' }
-#' @return  a list of 3 data.frames, a results data.frame (dplyr "tbl_df") with 11 columns:
+#' @return  A results data.frame (dplyr "tbl_df") with 11 columns:
 #' \itemize{
 #'  \item the name of the location ("location"),
 #'  \item the city it is in ("city"),
@@ -39,7 +45,7 @@
 #'  \item and finally an URL encoded version of the city name ("cityURL")
 #'  \item and of the location name ("locationURL").
 #' }.
-#' meta data.frame (dplyr "tbl_df") with 1 line and 5 columns:
+#' and two attributes, a meta data.frame (dplyr "tbl_df") with 1 line and 5 columns:
 #' \itemize{
 #' \item the API name ("name"),
 #' \item the license of the data ("license"),
@@ -48,10 +54,11 @@
 #' \item the limit on the number of results ("limit"),
 #' \item the number of results found on the platform for the query ("found")
 #' }
-#' last, a timestamp data.frame (dplyr "tbl_df") with the query time and the last time at which the data was modified on the platform.
+#' and a timestamp data.frame (dplyr "tbl_df") with the query time and the last time at which the data was modified on the platform.
 #' @export
 aq_latest <- function(country = NULL, city = NULL, location = NULL,# nolint
                    parameter = NULL, has_geo = NULL, limit = 100,
+                   latitude = NULL, longitude = NULL, radius = NULL,
                    page = 1) {
 
     ####################################################
@@ -63,12 +70,15 @@ aq_latest <- function(country = NULL, city = NULL, location = NULL,# nolint
                            parameter = parameter,
                            has_geo = has_geo,
                            limit = limit,
+                           latitude = latitude,
+                           longitude = longitude,
+                           radius = radius,
                            page = page)
 
     ####################################################
     # GET AND TRANSFORM RESULTS
     output <- getResults(urlAQ, argsList)
-    tableOfResults <- output$results
+    tableOfResults <- output
     # if no results
     if (nrow(tableOfResults) != 0){
 
@@ -81,13 +91,12 @@ aq_latest <- function(country = NULL, city = NULL, location = NULL,# nolint
 
     tableOfResults <- functionTime(tableOfResults,
                                    "lastUpdated")
-    if(ncol(tableOfResults) == 11){
-      names(tableOfResults)[4] <- "longitude"
-      names(tableOfResults)[5] <- "latitude"
-    }
+
+    names(tableOfResults) <- gsub("coordinates\\.", "", names(tableOfResults))
 
     }
-    return(list(results = tableOfResults,
-                meta = output$meta,
-                timestamp = output$timestamp))
+    attr(tableOfResults, "meta") <- attr(output, "meta")
+    attr(tableOfResults, "timestamp") <- attr(output, "timestamp")
+
+    return(tableOfResults)
 }